+++ lib/vquic/cf-ngtcp2.c
@@ -41,6 +41,7 @@
 #include "http1.h"
 #include "select.h"
 #include "transfer.h"
+#include "sockaddr.h"
 #include "bufref.h"
 #include "vquic/vquic.h"
 #include "vquic/vquic_int.h"
